Move all non-cmd folders into internal, split off nested packages
