+++ cmd/main.go
@@ -1,29 +1,29 @@
 package main
 
 import (
-	"github.com/skhanal5/clip-farmer/config"
-	"github.com/skhanal5/clip-farmer/internal/service"
+	"github.com/skhanal5/clip-farmer/internal/config"
+	"github.com/skhanal5/clip-farmer/internal/tiktokservice"
 )
 
 func main() {
 	configuration := config.NewConfig()
-	//
+
 	//Delegate the below portion in its own function: TwitchManager/Controller
-	//
-	//token := service.FetchTwitchOAuth(configuration)
+
+	//token := twitchservice.FetchTwitchOAuth(configuration)
 	//configuration.SetTwitchBearerToken(token)
 	//
-	//user := service.FetchUsers(configuration, "stableronaldo")
+	//user := twitchservice.FetchUsers(configuration, "stableronaldo")
 	//id := user.GetNthUser(0).Id
 	//
-	//resp := service.FetchUserClips(configuration, id)
+	//resp := twitchservice.FetchUserClips(configuration, id)
 	//err := downloader.DownloadClips(resp.Data)
 	//
 	//if err != nil {
 	//	panic(err)
 	//}
 
-	service.LoginIntoTargetUser(configuration)
-	//service.FetchTikTokOAuth(configuration)
+	tiktokservice.LoginIntoTargetUser(configuration)
+	//twitchservice.FetchTikTokOAuth(configuration)
 
 }
