+++ internal/tiktokmodel/tiktok_oauth_response.go
@@ -1,0 +1,11 @@
+package tiktokmodel
+
+type TikTokOAuthResponse struct {
+	AccessToken      string `json:"access_token"`
+	ExpiresIn        int    `json:"expires_in"`
+	OpenId           string `json:"open_id"`
+	RefreshExpiresIn int    `json:"refresh_expires_in"`
+	RefreshToken     string `json:"refresh_token"`
+	Scope            string `json:"scope"`
+	TokenType        string `json:"token_type"`
+}
