+++ internal/twitchmodel/twitch_oauth_response.go
@@ -1,0 +1,7 @@
+package twitchmodel
+
+type TwitchOAuthResponse struct {
+	AccessToken string `json:"access_token"`
+	ExpiresIn   int    `json:"expires_in"`
+	TokenType   string `json:"token_type"`
+}
